+++ src/local/vector.rs
@@ -15,7 +15,7 @@ use crate::{
 ///
 /// [`LocalVector`] also implements [Add](::std::ops::Add), [Sub](::std::ops::Sub),
 /// [Mul](::std::ops::Mul) and [Div](::std::ops::Div), among others.
-#[derive(PartialEq, Clone, Copy, Debug, Default)]
+#[derive(PartialEq, Clone, Copy, Debug)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))] // codecov:ignore:this
 pub struct LocalVector<O: FrameOrientation> {
     x: Length,